Fix program crash on startup
The simulator starts in front of a red light, causing it to occasionally
start in the APPROACH state, bypassing the CRUISE state. Since the
spline was only generated in the CRUISE state, this would result in a
crash when the velocity profile calculation was attempted before the
waypoints were initialized.

Regular path planning is likely a better approach in general as the
local plan needs evaluation (and possible modification) in real time.