Updates to decel logic
1. First point in deceleration trajectory is chosen to be the SPEED_LIMIT
value as opposed to current velocity. This fixed a behavior where the
car would be stuck in its tracks when starting in the APPROACH state
because it can't get going even though it still has space to the light.

2. Calculating time at each waypoint position in order to generate implied
velocity. This resulted in smoother deceleration from the correct
distance out (e.g. if we are targeting a decel of X m/s^2, then the slow
down will start X seconds from the hold position)

3. Update #2 (above) changed the velocity profile on decelerate, but the
car was unable to track the slow down, requiring a bump up in the brake
PID gain.

NOTE: The updated P value was chosen in order to match the P value for
the the throttle control, but could probably be tuned a bit better.